+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- methods for accessing parent node (`Node::parent`) and node object of attribute (`Attribute::nodeInfo`) - #53
+
+### Changed
+- `Location::name` will now return `/` for the root group instead of an empty string
 
 ## [3.0.0] - 2023-02-02
 ### Added